refactor(router): share response header validation
